tweak profile namespace for form model to match controller
