+++ app/controllers/jobseekers/profiles/qualified_teacher_status_controller.rb
@@ -1,27 +1,29 @@
 module Jobseekers
-  class Profiles::QualifiedTeacherStatusController < ProfilesController
-    def edit
-      @form = Profile::QualifiedTeacherStatusForm.new(profile.slice(*form_class.fields))
-    end
-
-    def update
-      @form = Profile::QualifiedTeacherStatusForm.new(form_params)
-      if @form.valid?
-        profile.update!(form_params)
-        redirect_to jobseekers_profile_qualified_teacher_status_path
-      else
-        render :edit
+  module Profiles
+    class QualifiedTeacherStatusController < Jobseekers::ProfilesController
+      def edit
+        @form = QualifiedTeacherStatusForm.new(profile.slice(*form_class.fields))
       end
-    end
 
-    private
+      def update
+        @form = QualifiedTeacherStatusForm.new(form_params)
+        if @form.valid?
+          profile.update!(form_params)
+          redirect_to jobseekers_profile_qualified_teacher_status_path
+        else
+          render :edit
+        end
+      end
 
-    def form_class
-      Profile::QualifiedTeacherStatusForm
-    end
+      private
 
-    def form_params
-      params.require(:jobseekers_profile_qualified_teacher_status_form).permit(form_class.fields)
+      def form_class
+        QualifiedTeacherStatusForm
+      end
+
+      def form_params
+        params.require(:jobseekers_profiles_qualified_teacher_status_form).permit(form_class.fields)
+      end
     end
   end
 end
