+++ app/views/jobseekers/base/_qts_fields.html.slim
@@ -7,14 +7,14 @@
     = f.govuk_text_area :qualified_teacher_status_details
   = f.govuk_radio_button :qualified_teacher_status, "on_track"
 
-= f.govuk_radio_buttons_fieldset :has_teacher_reference_number, hint: -> { tag.p(t("helpers.label.jobseekers_profile_qualified_teacher_status_form.teacher_reference_number_hint", link: govuk_link_to(t("helpers.label.jobseekers_profile_qualified_teacher_status_form.trn_link_text"), "https://find-a-lost-trn.education.gov.uk/start", target: "_blank")).html_safe) } do
+= f.govuk_radio_buttons_fieldset :has_teacher_reference_number, hint: -> { tag.p(t("helpers.label.jobseekers_profiles_qualified_teacher_status_form.teacher_reference_number_hint", link: govuk_link_to(t("helpers.label.jobseekers_profiles_qualified_teacher_status_form.trn_link_text"), "https://find-a-lost-trn.education.gov.uk/start", target: "_blank")).html_safe) } do
   = f.govuk_radio_button :has_teacher_reference_number, "yes", link_errors: true, label: { text: t("helpers.legend.jobseekers_job_application_professional_status_form.has_teacher_reference_number_options.yes") }
     = f.govuk_text_field :teacher_reference_number, label: { text: t("helpers.legend.jobseekers_job_application_professional_status_form.teacher_reference_number") }
   = f.govuk_radio_button :has_teacher_reference_number, "no", label: { text: t("helpers.legend.jobseekers_job_application_professional_status_form.has_teacher_reference_number_options.no") }
 
 = f.govuk_radio_buttons_fieldset :statutory_induction_complete,
-  legend: { text: t("helpers.label.jobseekers_profile_qualified_teacher_status_form.statutory_induction_complete") } do
+  legend: { text: t("helpers.label.jobseekers_profiles_qualified_teacher_status_form.statutory_induction_complete") } do
   = f.govuk_radio_button :statutory_induction_complete, "yes", link_errors: true
   = f.govuk_radio_button :statutory_induction_complete, "no" do
     = f.govuk_text_area :statutory_induction_complete_details,
-      label: -> { tag.p(t("helpers.label.jobseekers_profile_qualified_teacher_status_form.statutory_induction_complete_details", link: govuk_link_to(t("helpers.label.jobseekers_profile_qualified_teacher_status_form.statutory_induction_complete_details_hint_text"), "https://www.gov.uk/government/publications/induction-for-early-career-teachers-england", target: "_blank")).html_safe) }
+      label: -> { tag.p(t("helpers.label.jobseekers_profiles_qualified_teacher_status_form.statutory_induction_complete_details", link: govuk_link_to(t("helpers.label.jobseekers_profiles_qualified_teacher_status_form.statutory_induction_complete_details_hint_text"), "https://www.gov.uk/government/publications/induction-for-early-career-teachers-england", target: "_blank")).html_safe) }
