+++ app/views/jobseekers/profiles/qualified_teacher_status/_summary.html.slim
@@ -1,7 +1,7 @@
 = govuk_summary_list do |summary_list|
   - summary_list.with_row do |row|
     - row.with_key text: t(".status_title")
-    - row.with_value text: t("helpers.label.jobseekers_profile_qualified_teacher_status_form.qualified_teacher_status_options.#{profile.qualified_teacher_status}")
+    - row.with_value text: t("helpers.label.jobseekers_profiles_qualified_teacher_status_form.qualified_teacher_status_options.#{profile.qualified_teacher_status}")
     - row.with_action(text: t("buttons.change"), href: edit_jobseekers_profile_qualified_teacher_status_path, visually_hidden_text: "qualified teacher status")
 
   - if profile.qualified_teacher_status_year.present?
